requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
python-dotenv==1.2.1
requests==2.32.5
urllib3==2.5.0
gtfs-realtime-bindings==1.0.0
psycopg2-binary==2.9.9